backend/requirements.txt
fastapi
uvicorn
python-multipart
Pillow
numpy
tensorflow
